Features + bug fixes
Find GameObjects by name via the FindObject method in the Engine class. New grid query system demonstrated, this will be used to retain information about tiles. We still need to implement an actual matrix to hold all these tiles; right now it's just able to tell you the top-leftmost position of the cell you clicked in, the size of the cell you clicked in, but they're all equally sized anyway, and the position of the center of the cell you clicked in. This information will later be used for building. Also I was dumb but now I'm smarter and adding multiple Vectors of the same dimension will actually add them now; instead of just passing until the next subscriber call iteration. Also a bug was fixed where setting the position before the size of an object would not change the position at all. Another new feature: you no longer need an Update and a Start method in every GameObjects class, they're now optional! That's all I can think of off the top of my head.