Floating structure fix
Structures not connecting in any way to the main wall will now degenerate (destroy). This is using a breadth first search algorithm, failed to implement last night because I was tired and had very limited brain capacity. Anyway, uses breadth first search to find the first (if any) cell in a group of immediately-connected cells that is adjacent to the main wall structure when a wall is destroyed. If it cannot find any connection to the main wall then the group, being "floating" and disconnected, will be destroyed.